requirements.txt
cve_searchsploit
python-dotenv
requests
vulners
Levenshtein
pyyaml